scripting: Add type annotations to the API
Type annotations serve as an additional form of documentation and also
improve code clarity.

I opted to use 3.10+ style type hints because Python 3.9 is almost EOL,
and the old style is considered deprecated:
<https://peps.python.org/pep-0585/>

With this change, a clear check at the beginning of the module to was
added to ensure the minimum version is met.